feat(inputs.modbus): Add per-metric configuration style (#13507)
