+++ plugins/inputs/modbus/sample_general_begin.conf
@@ -49,4 +49,5 @@
   ## Define the configuration schema
   ##  |---register -- define fields per register type in the original style (only supports one slave ID)
   ##  |---request  -- define fields on a requests base
+  ##  |---metric   -- define fields on a metric base
   configuration_type = "register"
